+++ src/RadioList.php
@@ -2,35 +2,12 @@
 
   namespace Fiv\Form;
 
+  trigger_error('Deprecated. Use ' . \Fiv\Form\Element\RadioList::class, E_USER_DEPRECATED);
+
+
   /**
-   * Class TextArea
-   * Generate <input type="radio"> html tag
-   *
-   * @author  Ivan Shcherbak <dev@funivan.com>
-   * @package Fiv\Form
+   * @deprecated
    */
-  class RadioList extends \Fiv\Form\Element\Multiple {
+  class RadioList extends \Fiv\Form\Element\RadioList {
 
-    /**
-     * @return string
-     */
-    public function render() {
-      $html = '';
-      $currentValue = $this->getValue();
-      $attributes = $this->attributes;
-      $attributes['type'] = 'radio';
-
-      foreach ($this->options as $value => $text) {
-        $attributes['value'] = $value;
-        if ($currentValue == $value) {
-          $attributes['checked'] = 'checked';
-        } else {
-          unset($attributes['checked']);
-        }
-
-        $html .= '<label>' . static::tag('input', $attributes) . $text . '</label>';
-      }
-
-      return $html;
-    }
   }